reverted to more universal command line code

+++ src_analysis/work_hdfs/find_invalids.py
@@ -25,40 +25,8 @@ import prg_StatsInvPixel as stats
 DATES_FILE = 'clustering_invalid_filelists.txt'
 MOD02_DIRECTORY = '/home/koenig1/scratch-midway2/MOD02/clustering'
 MOD35_DIRECTORY = '/home/koenig1/scratch-midway2/MOD35/clustering'
-#OUTPUT_CSV = 'output07102019.csv'
-OUTPUT_CSV = 'test_for_dates.csv'
+OUTPUT_CSV = 'output07102019.csv'
 MAIN_DIR = '/home/koenig1/clouds/src_analysis/combined/test'
-
-
-def get_info(mod02_path):
-    '''
-    Searches for desired files and writes csv with each row being a MOD02
-    filename, patch number and number of invalid pixels in the patch
-
-    Inputs:
-        file(str): name of mod02 hdf file
-
-    Outputs:
-        None (writes csv)
-    '''
-    output_file = OUTPUT_CSV
-    file = mod02_path[-44:]
-    fillvalue_list, mod02_img = stats.gen_mod02_img(mod02_path)
-    mod02_patches = _gen_patches(mod02_img, normalization=False)
-    bname = os.path.basename(file)
-    date = bname[10:22]
-    # Finds corresponding MOD35
-    mod35 = glob.glob(MAIN_DIR + '/MOD35*.hdf')
-    if mod35:
-        mod35_path = mod35[0]
-        hdf_m35 = SD(mod35_path, SDC.READ)
-        clouds_mask_img = stats.gen_mod35_img(hdf_m35)
-        # Checks validity of pixels for each file and writes csv
-        stats.check_invalid_clouds2(output_file, file, mod02_patches,
-                                    clouds_mask_img, fillvalue_list, thres=0.3)
-    else:
-        print("No mod35 file downloaded for " + date)
-
 
 
 def get_invalid_info(dates_file=DATES_FILE, mod02_dir=MOD02_DIRECTORY,
@@ -110,6 +78,11 @@ def get_invalid_info2(file):
     Searches for desired files and writes csv with each row being a MOD02
     filename, patch number and number of invalid pixels in the patch
 
+    Note the difference between this function and the one above:
+    get_invalid_info() is designed to run in ipython3 while
+    get_invalid_info2() is optimized to run on the command line
+    using the multiprocessing library for task parallelization
+
     Inputs:
         file(str): name of mod02 hdf file
 
@@ -152,46 +125,32 @@ if __name__ == "__main__":
     
     #Initializes pooling process for parallelization
     pool = mp.Pool(processes=args.processors)
-
-
-    # Initializes output csv to be appended later
-    with open(args.outputfile, 'w') as csvfile:
-        outputwriter = csv.writer(csvfile, delimiter=',')
-        outputwriter.writerow(['filename', 'patch_no', 'inval_pixels'])
-    csvfile.close()
-    desired_files = glob.glob(MAIN_DIR + '/MOD02*.hdf')
-    pool.map(get_info, desired_files)
-    pool.close()
-    pool.join()
-
-
-
     # If output csv exits, assumes cutoff by RCC so deletes last entry
     # and appends only new dates
-    # if os.path.exists(args.outputfile):
-    #     print('Checking for completion')
-    #     completed = pd.read_csv(args.outputfile)
-    #     completed = completed[completed['filename'].notnull()]
-    #     last_file = completed.tail(1)['filename'].tolist()[0]
-    #     done = completed[completed['filename'] != last_file]
-    #     print('Writing updated csv w/ only completed MOD02 files')
-    #     done.to_csv(args.outputfile, index=False)
-    # else:
-    #     # Initializes output csv to be appended later
-    #     with open(args.outputfile, 'w') as csvfile:
-    #         outputwriter = csv.writer(csvfile, delimiter=',')
-    #         outputwriter.writerow(['filename', 'patch_no', 'inval_pixels'])
-    #     csvfile.close()
-    #     last_file = None
-    # # Finds name of desired MOD02 hdf files to be analyzed
-    # with open(args.dates_file, "r") as file:
-    #     dates = file.readlines()
-    # desired_files = dates[0].replace('hdf', 'hdf ').split()
-    # if last_file:
-    #     last_idx = desired_files.index(last_file)
-    #     desired_files = desired_files[last_idx:]
+    if os.path.exists(args.outputfile):
+        print('Checking for completion')
+        completed = pd.read_csv(args.outputfile)
+        completed = completed[completed['filename'].notnull()]
+        last_file = completed.tail(1)['filename'].tolist()[0]
+        done = completed[completed['filename'] != last_file]
+        print('Writing updated csv w/ only completed MOD02 files')
+        done.to_csv(args.outputfile, index=False)
+    else:
+        # Initializes output csv to be appended later
+        with open(args.outputfile, 'w') as csvfile:
+            outputwriter = csv.writer(csvfile, delimiter=',')
+            outputwriter.writerow(['filename', 'patch_no', 'inval_pixels'])
+        csvfile.close()
+        last_file = None
+    # Finds name of desired MOD02 hdf files to be analyzed
+    with open(args.dates_file, "r") as file:
+        dates = file.readlines()
+    desired_files = dates[0].replace('hdf', 'hdf ').split()
+    if last_file:
+        last_idx = desired_files.index(last_file)
+        desired_files = desired_files[last_idx:]
 
-    # pool.map(get_invalid_info2, desired_files)
-    # pool.close()
-    # pool.join()
+    pool.map(get_invalid_info2, desired_files)
+    pool.close()
+    pool.join()
